requirements.txt
langchain_openai
langgraph
openevals
openai
python-dotenv
pyright
mypy